+++ src/core/com/frostwire/search/WebSearchPerformer.java
@@ -39,7 +39,7 @@ public abstract class WebSearchPerformer extends AbstractSearchPerformer {
     private final int timeout;
     private final HttpClient client;
 
-    public WebSearchPerformer(int token, String keywords, int timeout) {
+    public WebSearchPerformer(long token, String keywords, int timeout) {
         super(token);
         this.keywords = keywords;
         this.timeout = timeout;
